backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
cfgv==3.4.0
chardet==5.2.0
coverage==7.4.1
decorator==5.1.1
distlib==0.3.8
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
identify==2.5.34
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
mysqlclient==2.2.4
nodeenv==1.8.0
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
virtualenv==20.25.0
wcwidth==0.2.13
